requirements.txt
alembic==1.5.4
algoliasearch==2.5.0
appdirs==1.4.4
black==20.8b1
cachetools==4.2.2
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
fastapi==0.63.0
google-auth==1.14.0
h11==0.12.0
idna==2.10
Mako==1.1.4
MarkupSafe==1.1.1
mypy-extensions==0.4.3
pathspec==0.8.1
psycopg2==2.8.6
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.7.3
PyJWT==1.7.1
python-dateutil==2.8.1
python-dotenv==0.17.1
python-editor==1.0.4
python-fity3==0.7
python-slugify==4.0.1
regex==2020.11.13
requests==2.25.1
rsa==4.0
six==1.15.0
SQLAlchemy==1.3.23
starlette==0.13.6
text-unidecode==1.3
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
urllib3==1.26.4
uvicorn==0.13.3
Werkzeug==2.0.1